Add files via upload
This data corresponds to loans associated with customers. We are provided with 500 observations (train +test) and 11 features.

The goal of the project is to predict whether a loan should be granted to a new customer or not. This analysis generate valuable insight for banks to consider which customers they should provide loan.

Variables such as loan status, paid off time, loan term,education, age,gender and several other features are provided