Create README.md
This data corresponds to loan payments associated with different customers. We are provided with 500 observations (train+test) and 11 features. Each observation is independent of the previous one.
The goal of the project is to predict whether a loan should be granted to a new customer or not based on historic data and generates valuable insights for the banks to consider which customer they should provide loans.

Variables such as loan status, loan term, due date, loan amount, education , gender and several other features are provided.

The project is done using Python.